Fix issue where object id's were not valid
Return type of openId() should be hid_t which is the object
identifier. If the value is returned as herr_t some typecasting
happens (perhaps) and the value is not valid anymore for further
actions like closeId(). Changing the return value type fixes this.

Somewhere in the library an invalid object id is still used when
calling closeId(). I didn't find after using a reasonable time for
searching it, so I'm putting in a workaround.